# swaggerui_src/swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/totalcost/{appName}": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Total cost for an application.",
        "operationId": "getTotalCost",
        "parameters": [
          {
            "type": "string",
            "description": "Works only with custom setup of cluster. Allow impersonation of test users (Required if Impersonate-Group is set)",
            "name": "Impersonate-User",
            "in": "header"
          },
          {
            "type": "string",
            "description": "Works only with custom setup of cluster. Allow impersonation of test group (Required if Impersonate-User is set)",
            "name": "Impersonate-Group",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Cost"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "definitions": {
    "ApplicationRegistration": {
      "description": "ApplicationRegistration describe an application",
      "type": "object",
      "required": [
        "name",
        "repository",
        "sharedSecret",
        "adGroups",
        "owner",
        "creator"
      ],
      "properties": {
        "adGroups": {
          "description": "AdGroups the groups that should be able to access the application",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "AdGroups"
        },
        "creator": {
          "description": "ApplicationOwner of the application (email). Can be a single person or a shared group email",
          "type": "string",
          "x-go-name": "Creator"
        },
        "machineUser": {
          "description": "MachineUser is on/off toggler of machine user for the application",
          "type": "boolean",
          "x-go-name": "MachineUser"
        },
        "name": {
          "description": "ApplicationName the unique name of the Radix application",
          "type": "string",
          "x-go-name": "Name",
          "example": "radix-canary-golang"
        },
        "owner": {
          "description": "ApplicationOwner of the application (email). Can be a single person or a shared group email",
          "type": "string",
          "x-go-name": "Owner"
        },
        "privateKey": {
          "description": "PrivateKey the private part of the deploy key set or returned\nafter successful application",
          "type": "string",
          "x-go-name": "PrivateKey"
        },
        "publicKey": {
          "description": "PublicKey the public part of the deploy key set or returned\nafter successful application",
          "type": "string",
          "x-go-name": "PublicKey"
        },
        "repository": {
          "description": "Repository the github repository",
          "type": "string",
          "x-go-name": "Repository",
          "example": "https://github.com/equinor/radix-canary-golang"
        },
        "sharedSecret": {
          "description": "SharedSecret the shared secret of the webhook",
          "type": "string",
          "x-go-name": "SharedSecret"
        },
        "wbs": {
          "description": "WBS information",
          "type": "string",
          "x-go-name": "WBS"
        }
      },
      "x-go-package": "github.com/equinor/radix-cost-allocation-api/api/cost/models"
    }
  }
}